# Project/backend/swagger.yaml
openapi: 3.0.0
info:
    title: 'HBL Banking API'
    version: 1.0.0
    description: 'A comprehensive banking API for HBL Clone application'
    contact:
        name: 'Muhammad Junaid Afzal'
        email: me@junaidafzal.dev
    license:
        name: ISC
        url: 'https://opensource.org/licenses/ISC'
servers:
    -
        url: 'http://localhost:8000'
        description: 'Development server'
    -
        url: 'https://your-production-url.com'
        description: 'Production server'
components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
    schemas:
...
        Loan:
            type: object
            properties: {_id: {type: string, description: 'Loan ID'}, userId: {type: string, description: 'User ID'}, loanType: {type: string, enum: [personal, home, car, business]}, amount: {type: number}, tenure: {type: integer}, interestRate: {type: number}, monthlyInstallment: {type: number}, status: {type: string, enum: [pending, approved, rejected, active, closed]}, applicationDate: {type: string, format: date-time}, approvedDate: {type: string, format: date-time, nullable: true}, disbursedDate: {type: string, format: date-time, nullable: true}, purpose: {type: string}, collateral: {type: string}, outstandingAmount: {type: number}, rejectionReason: {type: string, description: 'Reason for loan rejection'}, rejectedDate: {type: string, format: date-time, description: 'Date when loan was rejected'}, createdAt: {type: string, format: date-time}, updatedAt: {type: string, format: date-time}}
...
security:
    -
        bearerAuth: []
paths:
...
    /api/admin/loans:
        get:
            summary: 'Get all loan applications from all users (Admin only)'
            tags: [Admin]
            security: [{bearerAuth: []}]
            parameters: [{in: query, name: page, schema: {type: integer, default: 1}}, {in: query, name: limit, schema: {type: integer, default: 20}}, {in: query, name: status, schema: {type: string, enum: [pending, approved, rejected, active, closed]}}, {in: query, name: loanType, schema: {type: string, enum: [personal, home, car, business]}}, {in: query, name: minAmount, schema: {type: number}}, {in: query, name: maxAmount, schema: {type: number}}, {in: query, name: startDate, schema: {type: string, format: date}}, {in: query, name: endDate, schema: {type: string, format: date}}, {in: query, name: search, schema: {type: string}, description: 'Search by loan ID, user name, email, or purpose'}, {in: query, name: sortBy, schema: {type: string, default: applicationDate}}, {in: query, name: sortOrder, schema: {type: string, enum: [asc, desc], default: desc}}]
            responses: {'200': {description: 'Loans retrieved successfully', content: {application/json: {schema: {type: object, properties: {success: {type: boolean, example: true}, count: {type: number}, total: {type: number}, page: {type: number}, pages: {type: number}, data: {type: array, items: {allOf: [{$ref: '#/components/schemas/Loan'}, {type: object, properties: {user: {type: object, properties: {_id: {type: string}, firstName: {type: string}, lastName: {type: string}, email: {type: string}, customerNumber: {type: string}}}}}]}}}}}}}, '403': {description: 'Admin access required'}}
...
    /api/loans:
        get:
            summary: 'Get user''s loans'
            tags: [Loans]
            security: [{bearerAuth: []}]
            responses: {'200': {description: 'Loans retrieved successfully', content: {application/json: {schema: {type: object, properties: {success: {type: boolean, example: true}, count: {type: number, example: 2}, data: {type: array, items: {$ref: '#/components/schemas/Loan'}}}}}}}, '401': {description: Unauthorized}}
        post:
            summary: 'Apply for a loan'
            tags: [Loans]
            security: [{bearerAuth: []}]
            requestBody: {required: true, content: {application/json: {schema: {type: object, required: [loanType, amount, tenure, purpose], properties: {loanType: {type: string, enum: [personal, home, car, business], example: personal}, amount: {type: number, example: 500000}, tenure: {type: integer, example: 36}, purpose: {type: string, example: 'Home renovation'}, collateral: {type: string, example: 'Property documents'}}}}}}
            responses: {'201': {description: 'Loan application submitted successfully', content: {application/json: {schema: {type: object, properties: {success: {type: boolean, example: true}, message: {type: string, example: 'Loan application submitted successfully'}, data: {$ref: '#/components/schemas/Loan'}}}}}}, '400': {description: 'Validation error'}, '401': {description: Unauthorized}}
    '/api/loans/{id}':
        get:
            summary: 'Get loan details'
            tags: [Loans]
            security: [{bearerAuth: []}]
            parameters: [{in: path, name: id, required: true, schema: {type: string}, description: 'Loan ID'}]
            responses: {'200': {description: 'Loan details retrieved successfully', content: {application/json: {schema: {type: object, properties: {success: {type: boolean, example: true}, data: {$ref: '#/components/schemas/Loan'}}}}}}, '401': {description: Unauthorized}, '404': {description: 'Loan not found'}}
    /api/loans/calculate:
        post:
            summary: 'Calculate loan EMI'
            tags: [Loans]
            security: [{bearerAuth: []}]
            requestBody: {required: true, content: {application/json: {schema: {type: object, required: [amount, tenure, loanType], properties: {amount: {type: number, example: 500000}, tenure: {type: integer, example: 36}, loanType: {type: string, enum: [personal, home, car, business], example: personal}}}}}}
            responses: {'200': {description: 'EMI calculation successful', content: {application/json: {schema: {type: object, properties: {success: {type: boolean, example: true}, data: {type: object, properties: {loanAmount: {type: number, example: 500000}, tenure: {type: integer, example: 36}, interestRate: {type: number, example: 18.5}, monthlyInstallment: {type: number, example: 18000}, totalAmount: {type: number, example: 648000}, totalInterest: {type: number, example: 148000}}}}}}}}, '400': {description: 'Validation error'}, '401': {description: Unauthorized}}
...